8 refactor/add styles and loading (#11)
* add button spinner

* add error styles for Output

* refactor output block
