+++ src/components/Output/Output.module.scss
@@ -8,16 +8,42 @@
     height: 100%;
     gap: 15px;
 
+    .wrapperRow {
+        width: 100%;
+        display: flex;
+        flex-direction: row-reverse;
+        align-items: center;
+        justify-content: space-between;
+
+
+        >button {
+            width: 130px;
+            background-color: $cyan-op80;
+            border: 1px solid $white;
+            transition: 0.5s ease all;
+
+            &:hover {
+                background-color: $bright-dark-yellow;
+            }
+        }
+
+    }
+
     .outputBlock {
         width: inherit;
         height: inherit;
         background-color: $dark-grey;
-        color: $white;
+        color: $white-op20;
         padding: 30px 20px;
         border-radius: $radius-small;
 
         &_placeholder {
             color: $grey;
+        }
+
+        &_error {
+            color: $red;
+            border: 1px solid $red;
         }
     }
 }
